backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
Flask-Login
Flask-Env
Flask-RESTful
Flask-Mail
Flask-WTF
mysql-connector-python
python-dotenv
pandas
openai
Flask-JWT-Extended
Flask-API
Flask-Caching
Flask-Admin
Flask-Session
Flask-Accepts
requests
gunicorn
pytest
Flask-Testing
Celery

'''
$ pip list
Package                   Version
------------------------- -----------
alembic                   1.14.1
amqp                      5.3.1
aniso8601                 10.0.0
annotated-types           0.7.0
anyio                     4.8.0
asgiref                   3.8.1
attrs                     25.1.0
billiard                  4.2.1
blinker                   1.9.0
cachelib                  0.13.0
celery                    5.4.0
certifi                   2025.1.31
charset-normalizer        3.4.1
click                     8.1.8
click-didyoumean          0.3.1
click-plugins             1.1.1
click-repl                0.3.0
colorama                  0.4.6
distro                    1.9.0
Flask                     3.1.0
flask_accepts             1.0.1
Flask-Admin               1.6.1
Flask-API                 3.1
Flask-Caching             2.3.1
flask-cors                5.0.1
Flask-Env                 2.0.0
Flask-JWT-Extended        4.7.1
Flask-Login               0.6.3
Flask-Mail                0.10.0
Flask-Migrate             4.1.0
flask-mysql-connector     1.1.0
Flask-MySQLdb             2.0.0
Flask-RESTful             0.3.10
flask-restx               1.3.0
Flask-Session             0.8.0
Flask-SQLAlchemy          3.1.1
Flask-Testing             0.8.1
Flask-WTF                 1.2.2
greenlet                  3.1.1
gunicorn                  23.0.0
h11                       0.14.0
httpcore                  1.0.7
httpx                     0.28.1
idna                      3.10
importlib_resources       6.5.2
iniconfig                 2.0.0
itsdangerous              2.2.0
Jinja2                    3.1.5
jiter                     0.8.2
jsonschema                4.23.0
jsonschema-specifications 2024.10.1
kombu                     5.4.2
Mako                      1.3.9
MarkupSafe                3.0.2
marshmallow               3.26.1
msgspec                   0.19.0
mysql-connector-python    9.2.0
mysqlclient               2.2.7
numpy                     2.2.3
openai                    1.64.0
packaging                 24.2
pandas                    2.2.3
pip                       25.0.1
pluggy                    1.5.0
prompt_toolkit            3.0.50
pydantic                  2.10.6
pydantic_core             2.27.2
PyJWT                     2.10.1
pytest                    8.3.4
python-dateutil           2.9.0.post0
python-dotenv             1.0.1
pytz                      2025.1
referencing               0.36.2
requests                  2.32.3
rpds-py                   0.23.1
six                       1.17.0
sniffio                   1.3.1
SQLAlchemy                2.0.38
tqdm                      4.67.1
typing_extensions         4.12.2
tzdata                    2025.1
urllib3                   2.3.0
vine                      5.1.0
wcwidth                   0.2.13
Werkzeug                  3.1.3
WTForms                   3.2.1
'''